refactor: update text filter to new approach
