+++ koala/cogs/__init__.py
@@ -5,3 +5,4 @@ from .twitch_alert import TwitchAlert
 from .colour_role import ColourRole
 from .intro_cog import IntroCog
 from .react_for_role import ReactForRole
+from .text_filter import TextFilter
